requirements.txt
# Django.core
Django==4.2
djangorestframework==3.14.0

# Security
Pyjwt==2.8.0

# Tools
pillow==10.2.0

